gradio-app/requirements.txt
gradio
requests
pytest
python-dotenv
